feat: adopt no-implicit-any-catch
Errors in JavaScript can be any value.
Traditionally in TypeScript those were typed as `any`, which is a gap in
the type system and has caused real bugs more than once for us.
Since TS 3.9, it is also possible to mark errors as `unknown`, but it is
a voluntary annotation.

These lint rules enforce typing errors as `unknown`.

Errors can be cast easily using our utilities, such as  `asError()`.